feat: actualizar OrderController y ItemDto para manejar correctamente los IDs de productos y agregar producto a OrderItem

+++ business/orders/src/main/java/com/unir/d1001/orders/dto/ItemDto.java
@@ -1,4 +1,4 @@
 package com.unir.d1001.orders.dto;
 
-public record ItemDto(Long productId, int quantity) {
+public record ItemDto(int productId, int quantity) {
 }
