+++ business/orders/src/main/java/com/unir/d1001/orders/services/ProductService.java
@@ -14,7 +14,7 @@ public class ProductService {
     @Autowired
     AppHTTPClient appHTTPClient;
 
-    public Optional<ProductDto> getProductById(Long id) {
+    public Optional<ProductDto> getProductById(int id) {
         try {
             ProductDto product = appHTTPClient.getProductsClient().get()
                     .uri("/products/" + id)
